modified the datatype in Dto

+++ src/main/java/com/flmhospitals/dto/BedRequestDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class BedRequestDTO {
 	
-	private int bedNum;
+	private long bedNum;
 	
 	private long roomNumber;
